Enhancing control-env.sh file

+++ control-env.sh
@@ -1,24 +1,38 @@
 #!/usr/bin/env bash
 
 function stop {
-  echo "Stopping and removing containers"
+  echo "Stopping and removing containers..."
   docker-compose --project-name wios down
 }
 
 function cleanup {
-  echo "Removing volumes"
+  echo "Removing volumes..."
   docker-compose --project-name wios down --volumes --rmi all
-  echo "Deleting cache folders"
+  echo "Deleting cache folders..."
   rm -rf ./airflow/dags/__pycache__
   rm -rf ./facial_database/jupyter_notebooks/.ipynb_checkpoints
   rm -rf ./facial_database/python_scripts/__pycache__
-  echo "Deleting log folders"
+  echo "Deleting log folders..."
   rm -rf ./airflow/logs
 }
 
 function hard_cleanup {
-  echo "Deleting datasets"
+  read -p "HOLD ON THERE, YOU CINEPHILE! You are about to delete the '/datasets' folder and clean some unused stuff from Docker-Desktop.\
+  Are you sure you want to continue? [y/n] " -n 1 -r
+  echo
+  if [[ ! $REPLY =~ ^[Yy]$ ]]
+  then
+      echo "Wise choice. Don't forget to check '/control-env.sh' file to know more about this stage."
+      exit 1
+  fi
+  echo "Alright, here we go..."
+  echo "Deleting datasets..."
   rm -rf ./facial_database/datasets
+  echo "Cleaning Docker-Desktop..."
+  docker rm $(docker ps -f status=exited -aq)
+  docker rmi $(docker images -f "dangling=true" -q)
+  docker volume rm $(docker volume ls -qf dangling=true)
+
 }
 
 function initial_setup {
@@ -29,7 +43,7 @@ function initial_setup {
 }
 
 function start {
-  echo "Starting up"
+  echo "Starting up..."
   docker-compose --project-name wios up -d
 }
 
@@ -89,7 +103,7 @@ case $1 in
     ;;
 
   * )
-  printf "ERROR: Missing command\n  Usage: `basename $0` (start|initial_setup|stop|cleanup|token|logs|update)\n"
+  printf "ERROR: Missing command\n  Usage: `basename $0` (start|initial_setup|stop|cleanup|hard_cleanup|token|logs|update)\n"
   exit 1
     ;;
 esac
